components/file-watcher/requirements.txt
watchdog==2.3.1
stream-manager==1.1.1
awsiotsdk==1.12.2
cerberus==1.3.4